consolidate API code for core and plus, fix access/default change in sliceplus, shore up caller arguments in update
